api/requirements.txt
black==23.9.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
coverage==7.8.0
cryptography==44.0.2
dotenv==0.9.9
flake8==6.1.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
iniconfig==2.1.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.20.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
